requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
crispy-bootstrap5==0.7
cryptography==41.0.4
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==4.2.6
django-allauth==0.57.0
django-autoslug==1.9.9
django-browser-reload==1.12.0
django-cache-url==3.4.4
django-check-seo==0.6.4
django-cities-light==3.9.2
django-ckeditor==6.7.0
django-classy-tags==4.1.0
django-cms==3.11.4
django-cors-headers==4.3.0
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-extra-views==0.14.0
django-filer==3.1.0
django-formtools==2.4.1
django-js-asset==2.1.0
django-meta==2.4.0
django-phonenumber-field==7.2.0
django-polymorphic==3.1.0
django-sekizai==4.1.0
django-treebeard==4.7
djangocms-admin-style==3.2.6
djangocms-page-meta==1.4.3
djangorestframework==3.14.0
easy-thumbnails==2.8.5
environs==9.5.0
fontawesome-free==5.15.4
graphviz==0.20.1
gunicorn==21.2.0
idna==3.4
lxml==4.9.3
marshmallow==3.20.1
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.22
Pillow==10.0.1
progressbar2==4.2.0
psycopg==3.1.12
psycopg2-binary==2.9.9
pycparser==2.21
pydot==1.4.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
python-utils==3.8.1
python3-openid==3.2.0
pytz==2023.3.post1
pywatchman==1.4.1
reportlab==4.0.6
requests==2.31.0
requests-oauthlib==1.3.1
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.8.0
Unidecode==1.3.7
urllib3==2.0.6
webencodings==0.5.1
whitenoise==6.6.0
